refactor: moved facility locations from user state to util state(#31a0f22)
