+++ src/mixins/parseFileMixin.ts
@@ -6,12 +6,11 @@ export default {
   methods: {
     async parseCsv(file: any) {
       if (file) {
-        showToast(translate("File uploaded successfully"));
         store.dispatch('order/updateFileName', file.name);
-        
         const csvData = await parseCsv(file).then( res => {
           return  res;
         })
+        showToast(translate("File uploaded successfully"));
         return csvData;
       }
       showToast(translate("Something went wrong, Please try again"));
