+++ src/services/UtilService.ts
@@ -1,12 +1,22 @@
 import api from '@/api';
 
 const getFacilities= async (payload: any): Promise <any> => {
-    return api({
-      url: "/performFind",
-      method: "POST",
-      data: payload
-    });
-  }
+  return api({
+    url: "/performFind",
+    method: "POST",
+    data: payload
+  });
+}
+
+const getFacilityLocations = async (payload: any): Promise<any> => {
+  return api({
+    url: "/performFind",
+    method: "POST",
+    data: payload
+  })
+}
+
 export const UtilService = {
-    getFacilities
-  }
+    getFacilities,
+    getFacilityLocations
+}
