+++ src/store/modules/util/getters.ts
@@ -5,6 +5,9 @@ import RootState from '@/store/RootState'
 const getters: GetterTree <UtilState, RootState> = {
     getFacilities (state) {
         return state.facilities;
-    }
+    },
+    getFacilityLocationsByFacilityId: (state) => (facilityId: string) => {
+        return state.facilityLocationsByFacilityId[facilityId]
+    },
 }
 export default getters;
