+++ src/store/modules/util/mutation-types.ts
@@ -1,2 +1,3 @@
 export const SN_UTIL = 'util'
 export const UTIL_FACILITIES_UPDATED = SN_UTIL + '/FACILITIES_UPDATED'
+export const UTIL_FACILITY_LOCATIONS_BY_FACILITY_ID = SN_UTIL + '/FACILITY_LOCATIONS_BY_FACILITY_ID'
