+++ src/store/modules/util/mutations.ts
@@ -6,5 +6,10 @@ const mutations: MutationTree <UtilState> = {
     [types.UTIL_FACILITIES_UPDATED] (state, payload) {
         state.facilities = payload
     },
+    [types.UTIL_FACILITY_LOCATIONS_BY_FACILITY_ID] (state, facilityLocations) {
+        Object.keys(facilityLocations).map((facilityId: any) => {
+            state.facilityLocationsByFacilityId[facilityId] = facilityLocations[facilityId];
+        })
+    },
 }
 export default mutations;
